organise ir into separate files
