+++ nir/ir/common.py
@@ -1,0 +1,3 @@
+import numpy as np
+from .node import NIRNode
+from dataclasses import dataclass
