+++ nir/ir/flatten.py
@@ -1,0 +1,35 @@
+from .common import *
+from .typing import Types
+from .utils import parse_shape_argument, calc_flatten_output
+
+
+@dataclass(eq=False)
+class Flatten(NIRNode):
+    """Flatten node.
+
+    This node flattens its input tensor.
+    input_type must be a dict with one key: "input".
+    """
+
+    # Shape of input tensor (overrrides input_type from
+    # NIRNode to allow for non-keyword (positional) initialization)
+    input_type: Types
+    start_dim: int = 1  # First dimension to flatten
+    end_dim: int = -1  # Last dimension to flatten
+
+    def __post_init__(self):
+        self.input_type = parse_shape_argument(self.input_type, "input")
+        if self.input_type["input"] is None:
+            self.input_type = {"input": None}
+            self.output_type = {"output": None}
+        else:
+            self.output_type = {
+                "output": calc_flatten_output(
+                    self.input_type["input"], self.start_dim, self.end_dim
+                )
+            }
+            # make sure input and output shape are valid
+            if np.prod(self.input_type["input"]) != np.prod(self.output_type["output"]):
+                raise ValueError(
+                    "input and output shape must have same number of elements"
+                )
