+++ nir/ir/typing.py
@@ -1,0 +1,9 @@
+from typing import Dict, List, Tuple
+from .common import *
+
+# Nodes are uniquely named computational units
+Nodes = Dict[str, "NIRNode"]
+# Edges map one node id to another via the identity
+Edges = List[Tuple[str, str]]
+# Types is a dict mapping strings to tensor shapes
+Types = Dict[str, np.ndarray]
